Working on splitting code into separate modules (not working)
